# specs/001-ai-rag-eval-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI RAG-Based Document Evaluation & QnA Platform API
  description: API for document upload, RAG-based QnA, and automated evaluation
  version: 0.1.0
  contact:
    name: ELP Gemini Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: User login
      description: Authenticate user and create session
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: SecurePass123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      tags: [auth]
      summary: User logout
      description: End user session
      operationId: logout
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      description: Retrieve authenticated user information
      operationId: getCurrentUser
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Documents ====================
...
  /qna/{document_id}:
    post:
      tags: [qna]
      summary: Ask question
      description: Ask a question about the document content
      operationId: askQuestion
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  maxLength: 5000
                  example: What is the main conclusion in Chapter 3?
      responses:
        '200':
          description: Answer generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QAResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [qna]
      summary: Get QnA history
      description: Retrieve question-answer history for a document
      operationId: getQAHistory
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: QnA history
          content:
            application/json:
              schema:
                type: object
                properties:
                  qa_logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/QALogResponse'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Evaluation ====================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: integer
      description: Document ID
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        is_admin:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time

    # ==================== Document Schemas ====================
...
    QAResponse:
      type: object
      properties:
        question:
          type: string
          example: What is the main conclusion?
        answer:
          type: string
          example: The main conclusion is...
        model_name:
          type: string
          example: gemini-2.0-flash-exp
        latency_ms:
          type: integer
          example: 1250
        created_at:
          type: string
          format: date-time
    QALogResponse:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        model_name:
          type: string
        created_at:
          type: string
          format: date-time

    # ==================== Evaluation Schemas ====================
...
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Error description
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'